Move utilities to separate file and define main parsing flow
